+++ ldr.h
@@ -1,10 +1,10 @@
 #ifndef LDR_H
 #define LDR_H
 
-class LDR_sensor {  
+class LDR_sensor {
   public:
     LDR_sensor(int pin_num);
-    int readLDRLight();
+    float readLDRLight();
 
   private:
     int ldr_pin;
