+++ soil.h
@@ -1,10 +1,10 @@
 #ifndef SOIL_H
 #define SOIL_H
 
-class SOIL_sensor {  
+class SOIL_sensor {
   public:
     SOIL_sensor(int pin_num);
-    int readSOILMoist();
+    float readSOILMoist();
 
   private:
     int soil_pin;
